+++ Kernel/src/controlTools/SensorEvent.h
@@ -48,10 +48,10 @@ private:
 public:
 
   /** constructor with time value as a parameter
-   *  \param an unsigned long int
+   *  \param a double
    *  \param a string, type of Event
    */
-  SensorEvent(unsigned long int, const std::string&);
+  SensorEvent(double, const std::string&);
 
   /** destructor
   */
